requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pygbag==0.8.6
pygame-ce==2.4.1
jinja2==3.1.3
python-multipart==0.0.9
